pyproject.toml
[project]
name = "post-generator-agent"
version = "0.1.0"
description = "Twitter/X post generator agent powered by Claude Agent SDK"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.30",
    "python-dotenv>=1.2.1",
]
